# docs/swagger.yaml
basePath: /
definitions:
  handlers.CartInfo:
    properties:
      count:
        type: integer
      id:
        type: string
    type: object
  models.Artifact:
    properties:
      ID:
        type: string
      ImageURL:
        type: string
      Status:
        type: string
      description:
        type: string
      end_date:
        type: integer
      epoch:
        type: string
      name:
        type: string
      start_date:
        type: integer
      tpq:
        type: integer
    type: object
  models.TPQRequest:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      creatorID:
        type: integer
      excavation:
        type: string
      formedAt:
        type: string
      id:
        type: string
      moderatorID:
        type: integer
      result:
        description: Изменено на *int для nullable (пустое значение)
        type: integer
      status:
        type: string
    type: object
  models.TPQRequestItem:
    properties:
      artifact:
        $ref: '#/definitions/models.Artifact'
      artifactID:
        type: string
      comment:
        type: string
      requestID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing artifacts and TPQ requests.
  title: Chronus API
  version: "1.0"
paths:
...
  /api/tpq_requests:
    get:
      description: Get list of TPQ requests, filtered by user or all for moderator
      parameters:
      - description: Status filter
        in: query
        name: status
        type: string
      - description: From date
        in: query
        name: from_date
        type: string
      - description: To date
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TPQRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get TPQ requests
      tags:
      - tpq_requests
  /api/tpq_requests/{id}:
    delete:
      description: Delete a draft TPQ request (user required, own draft)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Cannot delete
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete TPQ request
      tags:
      - tpq_requests
    get:
      description: Get details of a specific TPQ request (user required, own or moderator)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Request not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get TPQ request by ID
      tags:
      - tpq_requests
    put:
      consumes:
      - application/json
      description: Update a TPQ request (user required, own draft)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TPQRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Request not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update TPQ request
      tags:
      - tpq_requests
  /api/tpq_requests/{id}/complete:
    put:
      description: Complete a formed TPQ request (moderator required)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "400":
          description: Cannot complete
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Complete TPQ request
      tags:
      - tpq_requests
  /api/tpq_requests/{id}/form:
    put:
      description: Form a draft TPQ request (user required)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "400":
          description: Cannot form
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Form TPQ request
      tags:
      - tpq_requests
  /api/tpq_requests/{id}/reject:
    put:
      description: Reject a formed TPQ request (moderator required)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "400":
          description: Cannot reject
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Reject TPQ request
      tags:
      - tpq_requests
  /api/tpq_requests/{request_id}/items/{artifact_id}:
    delete:
      description: Delete an item from TPQ request (user required, own draft)
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: string
      - description: Artifact ID
        in: path
        name: artifact_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete TPQ request item
      tags:
      - tpq_requests
    put:
      consumes:
      - application/json
      description: Update comment on TPQ request item (user required, own draft)
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: string
      - description: Artifact ID
        in: path
        name: artifact_id
        required: true
        type: string
      - description: Updated item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.TPQRequestItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequestItem'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update TPQ request item
      tags:
      - tpq_requests
  /api/tpq_requests/cart:
    get:
      description: Get current draft request info (user required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CartInfo'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get cart info
      tags:
      - tpq_requests
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"